requirements.txt
fastapi
uvicorn
tensorflow
opencv-python
ai-edge-litert
fastapi-mail
numpy
pymysql